+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add translation (panning) support too
 * Add `FvPath#destroy()` method
 * Add `FvPath#isConnectedTo(anchor_one, anchor_two)`
+* Make it possible to destroy anchors on the fly
 
 ## 0.1.0